Move split_dimensions and add dims_list in argument class
